RUGGUARD_LLM/requirement.txt
tweepy==4.14.0
python-dotenv==1.0.1
google-generativeai==0.5.4